updated layout and skeleton for discovery section
